+++ src/routes/home.css
@@ -77,53 +77,3 @@
     align-items: center;
     justify-content: space-between;
 }
-
-.reviewsArea{
-    display: flex;
-    overflow: hidden;
-    flex-direction: row;
-    justify-content: space-between;
-    margin-top: 5rem;
-    width: 100%;
-    height: 30vh;
-    white-space: nowrap;
-    background-color: #2525259e;
-    border-radius: 33px;
-    text-align: center;
-    backdrop-filter: blur(100px);
-}
-
-.reviewsContainer {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    background-color: black;
-    border-radius: 33px;
-    margin: 2%;
-}
-  
-.reviews{
-    flex-wrap: nowrap;
-    animation: move 15s infinite linear; /* set the time to what you want of course */
-    font-size: 20px;
-    padding-left:0px;
-    background-color: white;
-    border-radius: 33px;
-    width: 30vw;
-    height: 25vh;
-    margin: 30px;
-}
-
-.blurReview{
-    animation: blur-layer 10s infinite
-}
-
-@keyframes blur-layer {
-    0% { filter: blur(30px);}
-    1% { filter: blur(1px);}
-    2% { filter: blur(0px);}
-    50% { filter: blur(0px);}
-    98% {filter: blur(0px);}
-    99% { filter: blur(1px);}
-    100% { filter: blur(30px);}
-}
